+++ app/views/products/_suggestion.html.erb
@@ -1,6 +1,6 @@
-<% if suggestion = Spree::Config.searcher.suggest %>
+<% if suggestion = @searcher.suggest %>
   <p>
-    <%= t(:did_you_mean, :default => "Did you mean") %> 
+    <%= t(:did_you_mean, :default => "Did you mean") %>
     <%= link_to h(suggestion), :overwrite_params => { :keywords => suggestion } %>?
   </p>
 <% end %>
