+++ lib/spree_sphinx_search_hooks.rb
@@ -1,6 +1,4 @@
 class SphinxSearchHooks < Spree::ThemeSupport::HookListener
-
   insert_before :search_results, 'products/facets'
   insert_before :search_results, 'products/suggestion'
-  
 end
